FilesystemStoragetStrategy tests

+++ src/API/Heyer.Storage.API/Providers/Filesystem/FilesystemStorageStrategy.cs
@@ -23,6 +23,12 @@ internal class FilesystemStorageStrategy : IStorageStrategy
     public async Task StoreAsync(string key, Stream stream, CancellationToken cancellationToken = default)
     {
         var path = Path.Combine(_rootPath, key);
+
+        if (File.Exists(path))
+        {
+            throw new InvalidOperationException("File already exists.");
+        }
+        
         await using var fileStream = File.Create(path);
         stream.Seek(0, SeekOrigin.Begin);
         await stream.CopyToAsync(fileStream, cancellationToken);
@@ -38,6 +44,14 @@ internal class FilesystemStorageStrategy : IStorageStrategy
     public Task<Stream> GetAsync(string key, CancellationToken cancellationToken = default)
     {
         var path = Path.Combine(_rootPath, key);
-        return Task.FromResult<Stream>(File.OpenRead(path));
+        
+        try
+        {
+            return Task.FromResult<Stream>(File.OpenRead(path));
+        }
+        catch (FileNotFoundException e)
+        {
+            throw new FileNotFoundException("File not found.", e);
+        }
     }
 }
